特性(模块1): 特性(模块1): 工具(模块2): eslint ignore
aaa
BREAKING CHANGE:
aaa
ISSUES CLOSED: aaa

+++ .eslintrc.js
@@ -16,6 +16,6 @@ module.exports = {
   rules: {
     "react/jsx-filename-extension": "off",
     "prettier/prettier": "error",
-    "quotes": [1, "double"]
+    "quotes": [1, "single"]
   },
 };
